requirements.txt
python-telegram-bot==20.6
schedule==1.2.2
requests==2.31.0